requirements.txt
# requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9  # Для подключения к Postgres (binary-версия проще для Docker)